+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## 📦 unreleased
 
-* Breaking: Add sub-commands.
+* **Breaking:** Add sub-commands.
   This changes the search command from `pip-abandoned <path>` to `pip-abandoned search <path>`
+* **Breaking:** The exit code if one or more abandoned packages around found is now `9`
 * Add `set-token` command, allowing GitHub token to be stored in the system keyring
 * Add `--version` flag